pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.121.3",
    "greenlet>=3.2.4",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
]
